simple_retrain.py
#!/usr/bin/env python3
"""
Simple script to retrain the ML model with 26 features.
"""

import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor
import joblib
import json

def retrain_model():
    """Retrain model with 26 features."""
    print("Creating synthetic training data with 26 features...")
    
    # Generate 1000 synthetic property records
    n_samples = 1000
    np.random.seed(42)
    
    # Define feature columns (26 features matching feature_columns.json)
    feature_columns = [
        'bedrooms', 'bathrooms', 'square_feet', 'lot_size', 'rooms',
        'city_encoded', 'province_encoded', 'property_type_encoded',
        'year_built', 'current_year', 'current_month', 'dom', 'taxes',
        'policy_rate', 'prime_rate', 'mortgage_rate', 'inflation_rate',
        'unemployment_rate', 'exchange_rate', 'gdp_growth', 'interest_environment',
        'economic_momentum', 'affordability_pressure', 'property_affordability',
        'economic_sensitivity', 'market_timing'
    ]
    
    # Generate synthetic data
    data = {}
    
    # Basic property features
    data['bedrooms'] = np.random.randint(1, 6, n_samples)
    data['bathrooms'] = np.random.uniform(1, 4, n_samples)
    data['square_feet'] = np.random.randint(800, 4000, n_samples)
    data['lot_size'] = np.random.uniform(0.1, 2.0, n_samples)
    data['rooms'] = np.random.randint(3, 10, n_samples)
    data['year_built'] = np.random.randint(1950, 2023, n_samples)
    data['current_year'] = np.full(n_samples, 2025)
    data['current_month'] = np.random.randint(1, 13, n_samples)
    data['dom'] = np.random.randint(1, 180, n_samples)
    data['taxes'] = np.random.uniform(2000, 15000, n_samples)
    
    # Encoded categorical features
    data['city_encoded'] = np.random.randint(0, 5, n_samples)  # 5 cities
    data['province_encoded'] = np.random.randint(0, 3, n_samples)  # 3 provinces
    data['property_type_encoded'] = np.random.randint(0, 4, n_samples)  # 4 types
    
    # Economic indicators
    data['policy_rate'] = np.random.uniform(1.0, 6.0, n_samples)
    data['prime_rate'] = data['policy_rate'] + np.random.uniform(1.5, 2.5, n_samples)
    data['mortgage_rate'] = data['prime_rate'] + np.random.uniform(0.5, 2.0, n_samples)
    data['inflation_rate'] = np.random.uniform(1.0, 5.0, n_samples)
    data['unemployment_rate'] = np.random.uniform(3.0, 10.0, n_samples)
    data['exchange_rate'] = np.random.uniform(1.25, 1.40, n_samples)
    data['gdp_growth'] = np.random.uniform(-2.0, 4.0, n_samples)
    data['interest_environment'] = np.random.uniform(0, 1, n_samples)
    data['economic_momentum'] = np.random.uniform(-1, 1, n_samples)
    data['affordability_pressure'] = np.random.uniform(0, 1, n_samples)
    data['property_affordability'] = np.random.uniform(0, 1, n_samples)
    data['economic_sensitivity'] = np.random.uniform(0, 1, n_samples)
    data['market_timing'] = np.random.uniform(0, 1, n_samples)
    
    df = pd.DataFrame(data)
    
    # Create realistic price
    base_price = (
        df['square_feet'] * 250 +
        df['bedrooms'] * 15000 +
        df['bathrooms'] * 10000 +
        df['lot_size'] * 50000
    )
    
    # Add economic effects
    economic_effect = (
        (1 - df['interest_environment']) * 0.1 +
        df['economic_momentum'] * 0.05 +
        (1 - df['affordability_pressure']) * 0.1
    )
    
    df['price'] = base_price * (1 + economic_effect) + np.random.normal(0, 20000, n_samples)
    df['price'] = np.maximum(df['price'], 100000)
    
    # Prepare features and target
    X = df[feature_columns]
    y = df['price']
    
    print(f"Training data shape: {X.shape}")
    print(f"Number of features: {len(feature_columns)}")
    
    # Train the model
    print("Training RandomForest model...")
    model = RandomForestRegressor(n_estimators=100, random_state=42, max_depth=10)
    model.fit(X, y)
    
    # Save the model
    model_path = '/Users/efeobukohwo/Desktop/Nextproperty Real Estate/models/trained_models/property_price_model.pkl'
    joblib.dump(model, model_path)
    print(f"✅ Model saved to: {model_path}")
    
    # Verify feature columns file
    feature_path = '/Users/efeobukohwo/Desktop/Nextproperty Real Estate/models/model_artifacts/feature_columns.json'
    with open(feature_path, 'w') as f:
        json.dump(feature_columns, f, indent=2)
    print(f"✅ Feature columns saved to: {feature_path}")
    
    # Test the model
    print("Testing model...")
    sample_input = X.iloc[0:1]
    prediction = model.predict(sample_input)
    print(f"Sample prediction: ${prediction[0]:,.2f}")
    print(f"Sample input shape: {sample_input.shape}")
    
    print("✅ Model training completed successfully!")
    print("✅ Model now supports 26 features with economic integration!")

if __name__ == "__main__":
    retrain_model()
