requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Caching==2.1.0
Werkzeug==2.3.7

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.21

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
lightgbm==4.0.0
scipy==1.11.2

# Data Visualization
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2

# API Integration
requests==2.31.0
urllib3==2.0.4

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Validation and Security
marshmallow==3.20.1
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development and Testing (optional for production)
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0

# Background Tasks (optional)
celery==5.3.2
redis==4.6.0

# Monitoring and Logging
flask-monitor==1.2.0
